requirements.txt
# =============================================================================
# DEPRECATION WARNING:
#
# The file `requirements.txt` does not influence the package dependencies and
# will not be automatically created in the next version of PyScaffold (v4.x).
#
# Please have look at the docs for better alternatives
# (`Dependency Management` section).
# =============================================================================
#
# Add your pinned requirements so that they can be easily installed with:
# pip install -r requirements.txt
# Remember to also add them in setup.cfg but unpinned.
# Example:
# numpy==1.13.3
# scipy==1.0
#

setuptools~=49.6.0
pandas~=1.2.2
flask~=1.1.2
nltk~=3.4.5
swifter~=0.305
numpy~=1.19.5
scipy~=1.6.0
scikit-learn~=0.24.1
spacy~=2.3.5
mlxtend~=0.18.0
joblib~=1.0.1
Flask-PyMongo~=2.2.0
Scrapy~=2.4.1
selectorlib~=0.16.0
itemadapter~=0.2.0
python-dateutil~=2.7.3
crochet~=1.12.0